fixed nodemailer config
