+++ app.js
@@ -1,8 +1,7 @@
 const express = require('express');
 const logger = require('morgan');
 const cors = require('cors');
-const nodemailer = require('nodemailer');
-require("dotenv").config();
+
 
 const authRouter=require("./routes/api/auth-routes")
 
@@ -28,31 +27,5 @@ app.use((err, req, res, next) => {
   const {status=500, message="Server error" } = err;
   res.status(status).json({ message,})
 })
-
-const { KEY_NODEMAILER } = process.env;
-
-let transporter = nodemailer.createTransport({
-    service: 'gmail',
-    auth: {
-        user: 'mashamak175@gmail.com',
-      pass: { KEY_NODEMAILER }
-    }
-});
-
-
-let mailOptions = {
-    from: 'mashamak175@gmail.com',
-    to: 'piloge7795@gam1fy.com',
-    subject: 'Test Email',
-    html: '<p>Hello World!</p>'
-};
-
-transporter.sendMail(mailOptions, function(error, info){
-    if (error) {
-        console.log(error);
-    } else {
-        console.log('Email sent: ' + info.response);
-    }
-});
 
 module.exports = app;
