+++ helpers/index.js
@@ -1,5 +1,7 @@
 const HttpError = require("./HttpError");
+const sendMail =require("./sendemail")
 
 module.exports = {
     HttpError,
+    sendMail
 }
